Support closures as values for `::set()`
This allows us to `Memoize::set()` with a `Closure` as a value, resolving it when retrieving from memoization. If the key being retrieved is not itself a Closure, we will resolve the closures recursively from the key downwards.

Because `wp_cache_set()` should not store closures, when using the `WpCacheDriver`, any value that is set as a `Closure` will be stored in memory until the Closure is resolved.